refactor: extract symbol picker grid model and sync vendored emoji data semantically
Moves the grid's windowing/keyboard logic into useSymbolPickerIconGrid (no
ref reads during render, complexity under the cap), switches picker sizes to
bare token utilities, drops the unused static-copy plugin, and makes the
Emojibase mirror check compare parsed data so the repo formatter owns the
bytes.
