+++ web/package.json
@@ -40,7 +40,7 @@
         "@dagrejs/dagre": "^3.1.1",
         "@fontsource-variable/geist": "^5.3.0",
         "@fontsource-variable/jetbrains-mono": "^5.3.0",
-        "@lexical/react": "^0.49.0",
+        "@lexical/react": "^0.48.0",
         "@tailwindcss/vite": "^4.3.3",
         "@tanstack/react-query": "^5.102.4",
         "@tanstack/react-router": "^1.170.32",
@@ -50,7 +50,7 @@
         "@xyflow/react": "^12.11.5",
         "ai": "^7.0.79",
         "date-fns": "^4.4.0",
-        "lexical": "^0.49.0",
+        "lexical": "^0.48.0",
         "lucide-react": "^1.34.0",
         "motion": "^12.43.0",
         "openapi-fetch": "^0.17.0",
